Fix for currently released API (which needs sigs to be lower case hex).
MediaServer - ported to use new ExternalAPI library, but could use a lot of optimization. The code is fairly ugly, as it's built on top of the auto generated Intel UPNP project. The Intel library itself has been modified a bit (with help from some sources on the net as well as some reverse engineering I've done myself on the protocol), with most changes noted with comments.
The whole thing could use a rewrite :)